The coserver client code currently uses numeric UIDs to identify the user. This is a bad idea because 1) a user's UID can vary from machine to machine and 2) there is no such thing as a numeric UID in Windows.  Therefore, switch to using the user name instead.
On Windows, use a fake name ("UnknownWin32User") if GetUserName() fails or
the user name can not be translated to a multi-byte string in the current
locale.  Note that it is quite common for a Windows user name to contain
spaces and other non-alphanumeric characters.

On Unix, use a fake name ("UnknownUserX" where X is the user's UID) if
getpwuid() returns NULL.